+++ lib/src/zfa_cli.dart
@@ -7,7 +7,7 @@ import 'commands/config_command.dart';
 import 'commands/initialize_command.dart';
 import 'commands/entity_command.dart';
 
-const version = '2.3.2';
+const version = '2.4.0';
 
 Future<void> run(List<String> args) async {
   if (args.isEmpty) {